feat: Add Navigation Tests
This commit adds navigation tests to the project. The tests cover navigating to different screens in the app, including the Cant Eat screen, Like screen, and Recipe screen. Additionally, there is a test for navigating back to the Start screen. The tests use Compose UI testing APIs and are written in Kotlin.